requirements.txt
# Django and core dependencies
dj-database-url
psycopg2-binary

# Celery and Redis
celery
django-celery-beat

# Admin enhancements
django-admin-list-filter-dropdown
django-admin-searchable-dropdown

# UI and content
sweetify
django-tinymce

# API documentation
drf-spectacular

# Authentication and security
pyotp

# Testing
pytest
pytest-django
pytest-mock
freezegun

# Utilities
Faker
Pillow
python-decouple